to isolated domain entity definition, we extract the entites that mapping to JPA/DB Table to a single module, remove original packages entity defnition
