+++ hql-fabric/hql-fabric-example/src/main/java/com/hql/fabric/example/controller/SelectQueryController.java
@@ -1,7 +1,7 @@
 package com.hql.fabric.example.controller;
 
+import com.hql.fabric.domain.permission.User;
 import com.hql.fabric.example.dto.ApiResponse;
-import com.hql.fabric.example.entity.permission.User;
 import com.hql.fabric.example.loader.ExampleDatasetLoader;
 import com.hql.fabric.persistence.query.builder.HqlQueryBuilder;
 import com.hql.fabric.persistence.service.impl.HqlQueryService;
