--- hql-fabric/hql-fabric-example/src/main/java/com/hql/fabric/example/entity/permission/package-info.java
@@ -1,11 +1,0 @@
-package com.hql.fabric.example.entity.permission;
-
-/**
- * Permission Entities Mapping Relationship
- * <p>
- * User <-> UserProfile => 1:1
- * User <-> Role => M:N
- * Role <-> Permission => M:N
- * User <-> Group => M:N
- * Group <-> Role => M: N
- */
